Restructured project
This is part one of changing `breakable_toys` repo to be only for Go,
and only for one project. This will be more correct with the Go
workspace concept. Also it just feels more correct and will be less
annoying to wade deep into the project before being able to do work.