Tried to decouple serializer from details of JSON
Attempted to decouple the concept of a serializer from the concept of a
JSON formatter. Formatting JSON seems to be a separate responsibility
from mapping it to a type: the details of JSON syntax are at a different
level of abstraction than its overall structure.

I was trying to be all spiffy and cool with closures, too... But it is
not working. I am baffled. Gilli, help!!